requirements.txt
Flask==2.3.2
pytesseract==0.3.10
opencv-python-headless==4.8.0
numpy==1.25.0
python-docx==0.8.11
pdf2image==1.16.3
Werkzeug==2.3.6
